@@ -1,62 +1,76 @@
 {
   "name": "dragondropweb",
-  "version": "0.0.0",
-  "license": "MIT",
+  "version": "1.0.0",
+  "license": "UNLICENSED",
   "scripts": {
     "ng": "ng",
     "start": "concurrently --names \"CLIENT,SERVER\"  -c \"bgBlue.bold,bgGreen.bold\" \"yarn run client-proxy\" \"yarn run server\"",
+    "start-nocheck": "concurrently --names \"CLIENT,SERVER\" -c \"bgBlue.bold,bgGreen.bold\" \"yarn run client-proxy-nocheck\" \"yarn run server\"",
     "start-client": "ng serve",
     "build-client": "ng build",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "client-proxy": "ng serve --proxy proxy.conf.json",
+    "client-proxy": "ng serve --host 0.0.0.0 --proxy-config proxy.conf.json",
+    "client-proxy-nocheck": "ng serve --proxy proxy.conf.json --host 127.0.0.1 --disable-host-check",
     "server": "nodemon server.js"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.4",
-    "@angular/cdk": "^5.2.1",
-    "@angular/cli": "^1.6.7",
-    "@angular/common": "^5.2.4",
-    "@angular/compiler": "^5.2.4",
-    "@angular/core": "^5.2.4",
-    "@angular/forms": "^5.2.4",
-    "@angular/http": "^5.2.4",
-    "@angular/material": "^5.2.1",
-    "@angular/platform-browser": "^5.2.4",
-    "@angular/platform-browser-dynamic": "^5.2.4",
-    "@angular/router": "^5.2.4",
+    "@angular/animations": "^6.0.2",
+    "@angular/cdk": "^6.0.2",
+    "@angular/cli": "^6.0.1",
+    "@angular/common": "^6.0.2",
+    "@angular/compiler": "^6.0.2",
+    "@angular/core": "^6.0.2",
+    "@angular/forms": "^6.0.2",
+    "@angular/http": "^6.0.2",
+    "@angular/material": "^6.0.2",
+    "@angular/platform-browser": "^6.0.2",
+    "@angular/platform-browser-dynamic": "^6.0.2",
+    "@angular/router": "^6.0.2",
+    "@auth0/angular-jwt": "2.0.0",
     "angular-froala-wysiwyg": "^2.7.5",
+    "argon2": "^0.19.2",
     "auth0-js": "^9.3.0",
     "body-parser": "^1.18.2",
     "codemirror": "^5.35.0",
     "core-js": "^2.4.1",
     "express": "^4.16.2",
     "express-formidable": "^1.0.0",
+    "express-jwt": "^5.3.1",
     "font-awesome": "^4.7.0",
     "froala-editor": "^2.7.5",
-    "fs-extra": "^5.0.0",
+    "fs-extra": "^6.0.1",
     "hammerjs": "^2.0.8",
+    "helmet": "^3.12.0",
+    "joi": "^13.2.0",
+    "jsonwebtoken": "^8.2.1",
+    "minio": "^5.0.0",
     "mongoose": "^5.0.5",
     "multer": "^1.3.0",
     "ngx-bootstrap": "^2.0.2",
     "ngx-editor": "^3.3.0-rc.7",
-    "ngx-quill": "^2.1.1",
+    "ngx-quill": "^3.1.0",
     "primeng": "^5.2.0",
     "quill": "^1.3.5",
-    "quill-delta-to-html": "^0.7.2",
-    "rxjs": "^5.4.2",
+    "quill-delta-to-html": "^0.8.2",
+    "rxjs": "^6.1.0",
+    "rxjs-compat": "^6.1.0",
     "semver": "^5.5.0",
+    "typescript": ">=2.7.0 <2.8.0",
     "zone.js": "^0.8.14"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "^5.2.4",
-    "@angular/language-service": "^5.2.4",
+    "@angular-devkit/build-angular": "~0.6.1",
+    "@angular/compiler-cli": "^6.0.2",
+    "@angular/language-service": "^6.0.2",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~9.4.6",
-    "codelyzer": "~4.2.1",
+    "@types/minio": "^5.1.0",
+    "@types/node": "~10.1.0",
+    "angular-cli-ghpages": "^0.5.3",
+    "codelyzer": "~4.3.0",
     "concurrently": "^3.5.1",
     "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
@@ -65,10 +79,10 @@
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine-html-reporter": "^1.1.0",
+    "nodemon": "^1.17.3",
     "protractor": "~5.3.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "^2.7.1"
+    "ts-node": "~6.0.3",
+    "tslint": "~5.10.0"
   }
 }
