@@ -10,17 +10,25 @@ const bodyParser = require('body-parser');
 const path = require('path');
 const apiRoutes = require('./server/routes/api.route');
 const downloadRoutes = require('./server/routes/download.route');
-const mongoURI = 'mongodb://localhost/dragondrop';
-mongoose.connect(mongoURI, {server: {socketOptions: {keepAlive: 1}}})
+const helmet = require('helmet');
+const config = require('./server/config/config.dev');
+
+
+mongoose.connect(config.mongo.host, { server: { socketOptions: { keepAlive: 1 } } })
   .then(() =>{
     console.log('Connected to dragondrop db at localhost');
   })
   .catch(err => console.error(err));
 
-mongoose.set('debug', (collectionName, method, query, doc) => {
-  console.log(`${collectionName}.${method}`, doc);
-});
+// print mongoose logs in dev env
+if (config.mongooseDebug) {
+  console.log('Debug Moongoose');
+  mongoose.set('debug', (collectionName, method, query, doc) => {
+    console.log(`${collectionName}.${method}`, doc);
+  });
+}
 
+app.use(helmet());
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({extended: true}));
 
@@ -30,8 +38,12 @@ app.use('/api', apiRoutes);
 
 app.use('/download', downloadRoutes);
 
+app.use('/docs', express.static(path.join(__dirname, 'server/static/docs')));
+
+app.use('/test', express.static(path.join(__dirname, 'server/static/docs')));
+
 app.get('*', (req, res) => {
   res.sendFile(path.join(__dirname, 'dist/index.html'));
 });
 
-app.listen(3000, () => console.log('listening on port 3000'));
+app.listen(config.port, () => console.log(`listening on port ${config.port}`));
