@@ -12,6 +12,14 @@ import {DownloadLatestComponent} from './download-latest/download-latest.compone
 import {VersionListComponent} from './version-list/version-list.component';
 import {CallbackComponent} from './callback/callback.component';
 import {AuthGuard} from './auth/auth.guard';
+import {LoginComponent} from './login/login.component';
+import {DownloadInstructionsComponent} from './download-instructions/download-instructions.component';
+import {AdminPanelComponent} from './admin-panel/admin-panel.component';
+import {AdminUserListComponent} from './admin-user-list/admin-user-list.component';
+import {AdminPermissionsComponent} from './admin-permissions/admin-permissions.component';
+import {AdminUserTableComponent} from './admin-user-table/admin-user-table.component';
+import {AboutComponent} from "./about/about.component";
+import {DragondropDocumentationComponent} from "./dragondrop-documentation/dragondrop-documentation.component";
 
 const routes: Routes = [{
   path: '',
@@ -20,12 +28,25 @@ const routes: Routes = [{
   path: 'home',
   redirectTo: ''
 }, {
-  path: 'releases/:channel',
+  path: 'login',
+  component: LoginComponent
+}, {
+  path: 'release/:id',
   component: ReleaseComponent
 }, {
   path: 'admin',
   canActivate: [AuthGuard],
-  component: ReleaseListComponent
+  component: AdminPanelComponent,
+  children: [{
+    path: 'releases',
+    component: ReleaseListComponent
+  }, {
+    path: 'users',
+    component: AdminUserTableComponent
+  }, {
+    path: 'permissions',
+    component: AdminPermissionsComponent
+  }]
 }, {
   path: 'release',
   component: DownloadLatestComponent
@@ -36,12 +57,21 @@ const routes: Routes = [{
   path: 'callback',
   component: CallbackComponent
 }, {
+  path: 'instructions',
+  component: DownloadInstructionsComponent
+},{
+  path: 'documentation',
+  component: DragondropDocumentationComponent
+}, {
+  path: 'about',
+  component: AboutComponent
+}, {
   path: '**',
   component: PageNotFoundComponent
 }];
 
 @NgModule({
-  imports: [RouterModule.forRoot(routes, {enableTracing: true})],
+  imports: [RouterModule.forRoot(routes)],
   exports: [RouterModule]
 })
 export class AppRoutingModule {
