@@ -12,7 +12,7 @@ import {AuthService} from './auth/auth.service';
 })
 export class AppComponent {
 
-  constructor(private auth: AuthService) {
+  constructor(public auth: AuthService) {
     auth.handleAuthentication();
   }
 }
