@@ -5,6 +5,7 @@
 import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute} from '@angular/router';
 import {ReleaseService} from '../release.service';
+import {Release} from './release.model';
 
 const validBranches = ['stable', 'beta', 'alpha'];
 
@@ -16,28 +17,19 @@ const validBranches = ['stable', 'beta', 'alpha'];
 export class ReleaseComponent implements OnInit {
   error = false;
   errorMsg: String;
+  release?: Release = null;
 
   constructor(private route: ActivatedRoute,
-              private releaseService: ReleaseService) {
+              public releaseService: ReleaseService) {
   }
 
   ngOnInit() {
     this.errorMsg = '';
-    const branch = this.route.snapshot.paramMap.get('channel');
-    if (!validBranches.includes(branch)) {
-      this.error = true;
-      this.errorMsg += 'Invalid Release Channel';
-    }
+    const id = this.route.snapshot.paramMap.get('id');
+    console.log('id', id);
 
-    // const files = this.releaseService
-    //   .getLatestReleaseFilesForPlatform('windows');
-
-    const files = [];
-    if (files.length === 0) {
-      this.error = true;
-      this.errorMsg += 'No files available for';
-    }
-    console.log(branch);
+    this.releaseService.getRelease(id).subscribe(release => {
+      this.release = release;
+    });
   }
-
 }
