@@ -8,5 +8,6 @@
 // The list of which env maps to which file can be found in `.angular-cli.json`.
 
 export const environment = {
-  production: false
+  production: false,
+  authCallback: 'http://localhost:4200/callback'
 };
